% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateSource.R
\name{CreateSource}
\alias{CreateSource}
\title{Creates an interaction matrix (inter) and a dataframe with species parameters needed for the dynamical model.}
\usage{
CreateSource(inter, descSp, setPref = TRUE)
}
\value{
a list containing a dataframe with species parameters and a matrix with species interaction preferences
}
\description{
Creates an interaction matrix (inter) and a dataframe with species parameters needed for the dynamical model.
}
