% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakePhylo.R
\name{makePhylo}
\alias{makePhylo}
\title{Creates a dataframe with species scores on the 2 first axes of the Pcoa of their taxonomy.}
\usage{
makePhylo(NewTaxo, MakePhyloO = TRUE)
}
\value{
a list with 2 dataframes with species as rows, their names in the row.names, and their scores on the 2 first axes of the Pcoa of their taxonomy. YourData contains your species scores and ParamData contains parametrization species scores.
}
\description{
Creates a dataframe with species scores on the 2 first axes of the Pcoa of their taxonomy.
}
