% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictInter.R
\name{predictInter}
\alias{predictInter}
\title{Creates a dataframe species interactions}
\usage{
predictInter(descSp, axes)
}
\value{
a matrix containing in order a column with consumers name, a column with resources name, and a column with the interactions between them (1 or 0).
}
\description{
Creates a dataframe species interactions
}
