% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunModel.R
\name{the name you want to give to your simulation (character)}
\alias{the name you want to give to your simulation (character)}
\alias{RunModel}
\title{Creates a dataframe species interactions}
\usage{
RunModel(param, pref, name, vArg)
}
\arguments{
\item{the}{name of a dataframe (ex:'param.txt') with species parameters created by 'CreateSource'}
}
\value{
a .txt with species biomass as a function of time
}
\description{
Creates a dataframe species interactions
}
